+++ src/javascript/9000-main.js
@@ -1,4 +1,4 @@
-v.version.set(0,1,4);
+v.version.set(0,1,5);
 
 /*
 Sample code demonstrating the node system.
@@ -9,50 +9,37 @@ v.knowledge = [];
 
 s.makeNode('root');
 
-    f.build('remember', 'Action');
-    v.remember.setText(
-        'Calmly stand still and try to remember something ' +
-        'about how to survive a bear encounter'
-    );
-    v.root.push('remember');
+v.root.addLink(
+    'Calmly stand still and try to remember something ' +
+    'about how to survive a bear encounter',
 
-        f.build('rememberSpray', 'Outcome');
-        v.rememberSpray.carryOut = function() {
-            v.XP += 1;
-            v.knowledge.push('spray');
-            v.bearCharges.load();
-        }
-        v.remember.push('rememberSpray');
+    'bearCharges',
 
-    f.build('runAway', 'Action');
-    v.runAway.setText(
-        'Turn around and run down the hall'
-    );
-    v.root.push('runAway');
+    function() {
+        v.XP += 1;
+        v.knowledge.push('spray');
+    }
+);
 
-        f.build('tooSlow', 'Outcome');
-        v.tooSlow.carryOut = function() {
-            v.XP += 1;
-            v.bearChases.setTextSubs(['down the hall']);
-            v.bearChases.load();
-        }
-        v.runAway.push('tooSlow');
+v.root.addLink(
+    'Turn around and run down the hall',
 
-    f.build('runToLounge', 'Action');
-    v.runToLounge.setText(
-        'Run to the faculty lounge'
-    );
-    v.root.push('runToLounge');
+    'bearChases',
 
-        f.build('tooSlow2', 'Outcome');
-        v.tooSlow2.carryOut = function() {
-            v.XP += 1;
-            v.bearChases.setTextSubs(['toward the lounge']);
-            v.bearChases.load();
-        }
-        v.runToLounge.push('tooSlow2');
-
-s.makeNode('bearCharges');
-
-s.makeNode('bearChases');
+    function() {
+        v.XP += 1;
+        v.bearChases.setTextSubs(['down the hall']);
+    }
+);
 v.bearChases.setSubCount(1);
+
+v.root.addLink(
+    'Run to the faculty lounge',
+
+    'bearChases',
+
+    function() {
+        v.XP += 1;
+        v.bearChases.setTextSubs(['toward the lounge']);
+    }
+);